Update the mirror cover open/close sequence
to match that suggested by Tekniker.
Also split the sequence into two separate batches of commands:
one to move the mirror cover locks, the other to move the
mirror covers. That way a subsystem is less likely to be left on
if something goes wrong.

Rename the MirrorCoverLocksMoveAll "lock" paramter to "deploy"
for clarity.