Overhaul the way the command port is set
When talking to real Tekniker code the source ID must be set
correctly, else replies go to the wrong port.
The default source ID is set when the replies module is imported.
So base that default on CSC_COMMAND_PORT in constants.py
and do not allow that value to be changed on the command line.
(For unit tests the mock control port can still be any value,
since the mock controller works with any source ID).