MTMountCsc: output the target event in response to the moveToTarget command
Also add a wait_done flag to send_command.
Use this wait_done to send an IN_PROGRESS ack in response to the moveToTarget command
and to send the target event after moveToTarget has been acknowledged
(and thus is likely to succeed) but before the move is finished.
Also use wait_done to send an IN_PROGRESS ack for the mirror cover open/close commands.