mock device improvements
Add mock.DeployableDevice, a subclass of mock.PointToPointDevice
used for mirror covers and mirror cover locks (and deployable platform,
should we ever need that--not until the CSC is allowed to control it).
This has deploy and retract commands and a motion_state method.
Note that mirror cover locks and mirror covers have slightly different
deploy and retract commands, so this class supports both.

mock.AxisDevice:

* Add motion_state method. This required adding metadata
  for point-to-point moves.