In ``enums.py``, add new CSC error codes for when the CSC fails to enable or disable devices.
These error codes are trown when loading new settings.
During this procedure the CSC needs to disable and then enable the devices (elevation and azimuth) in order to load the new settings.
If during this operation the TMA fails to bring the devices down or up, the CSC will go to FAULT and trow one of these errors, depending which operation it is executing.