In MTMountCSC, fix an issue with the connect method that would return after sending the CSC to fault when there is an error.
This was causing a logic error where the CSC would go to fault when executing the start command, but then proceed to execute the start command normally, as there was no interruption/exception raised.

This update will make sure the CSC will go to fault and interrupt the state transition from continuing.