In MTMountCSC, update do_trackTarget to allow skipping commands that had timeout internally on the TMA side and to log the command history when track target command fails.
Sometimes the TMA software will have timeouts when sending track demands internally. We noticed that, most times it is ok to skip one or two subsequent failures. This change you ensure that the CSC will ignore these particular issues, so long as they happen at most a max_subsequent_failed_track_target number of times. In any case, the CSC will log the command history when this happens so we can track down what was going on.